pyproject.toml
[project]
name = "imagestats-api"
version = "0.0.1"
description = "Extract statistics (exif and color data) from images"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litserve>=0.2.10",
    "numpy>=2.2.6",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
]
